fix(matte): dedupe hex parsing, fix AppTheme literals, align file structure
Reuse TextStyle.RGBA(hex:) instead of reimplementing hex parsing (was
rejecting 8-char alpha hex). Replace hardcoded toolbar/sheet sizing with
AppTheme.IconSize.mdLg and new AppTheme.Matte constants. Split Matte.swift
into Models/Matte.swift (rendering) and EditorViewModel+Matte.swift
(asset creation), matching the rest of the codebase's media-asset layout.
